+++ varasto/forms.py
@@ -1,11 +1,23 @@
 from django import forms
-from django.contrib.auth.forms import UserCreationForm, AuthenticationForm
-from django.contrib.auth.models import User
-from django.core.exceptions import ValidationError
+from django.db import models
+from .models import CustomUser, Category
+from django.forms import ModelForm, widgets, TextInput, CheckboxInput
 
 
 
+class CustomUserForm(ModelForm):
+    class Meta:
+        model = CustomUser
+        fields = ['username', 'password']
+        widgets = {
+            'username': TextInput(attrs={
+                'class': 'form-control left',
+                'placeholder': 'login',
+            }),
+            'password': TextInput(attrs={
+                'class': 'form-control left',
+                'placeholder': 'password',
+            }),
 
-class LoginUserForm(AuthenticationForm):
-    username = forms.CharField(label='Логин', widget=forms.TextInput(attrs={'class': 'form-input'}))
-    password = forms.CharField(label='Пароль', widget=forms.PasswordInput(attrs={'class': 'form-input'}))
+        }
+
