+++ varasto/templates/varasto/index.html
@@ -1,0 +1,16 @@
+{% extends "varasto/base.html" %}
+
+{% load static %}
+
+{% block title %}Login{% endblock %}
+
+{% block content %}
+
+<div>
+    <h1>NEW APP!!!</h1>
+    <a href="{% url 'logout' %}"><button type="button">Logout</button></a>
+</div>
+
+
+
+{% endblock %}
