+++ varasto/templates/varasto/login.html
@@ -1,0 +1,21 @@
+{% extends "varasto/base.html" %}
+
+{% load static %}
+
+{% block title %}Login{% endblock %}
+
+{% block content %}
+
+<div>
+    <form method="post" autocomplete="off" enctype="multipart/form-data">
+        {% csrf_token %}
+        {{form.as_p}}
+
+
+        <button type="submit">Login</button>
+    </form>
+</div>
+
+
+
+{% endblock %}
